% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd.mnl.R
\name{rd.mnl}
\alias{rd.mnl}
\title{Regression Discontinuity (RD) Design for Categorical Outcomes}
\usage{
rd.mnl(DAT, c, H0_t, H0_R, H0_r, level)
}
\arguments{
\item{DAT}{The data, n by (J+2). n is the sample size and J+1 is the number of categories for the 
outcome. The first column of DAT is the running variable, the rest J+1 columns are binary dummies 
for each of the J+1 categories (each row sums to one). [If you are familiar with MATLAB, 
the last J+1 columns follows the format in the command mnrfit] The matrix DAT should have 
been sorted (ascendingly) according to the first column.}

\item{c}{The cutoff}

\item{H0_t}{J by 1. The null hypothesis when you perform a t-test.}

\item{H0_R}{q by J, where q is number of restrictions. H0_R and H0_r are restriction
constants as in the null hypothesis when you perform a Wald-test. E.g. H0_R=I_J 
if we want a joint significance test, H0_R=[1,-1] if we want a test of equivalence for the
treatment effects across two categories.}

\item{H0_r}{q by 1. See above.}

\item{level}{Confidence interval level. e.g. level=0.95.}
}
\value{
The result is a list, which contains following results.

h_opt: Optimal bandwidth (a scalar). 

ATE: Point estimates (J by 1). 

ttest: (Standard/Non-robust) t-test statistics and p-values (J by 2). 

ttest_rob: Robust t-test statistics and p-values. 


ci: (Standard/Non-robust) Confidence interval. 

ci_rob: Robust confidence interval. 

Waldtest: (Standard/Non-robust) Wald test statistic and the p-value (1 by 2). 

Waldtest_rob: Robust Wald statistic.
}
\description{
This function gives the optimal bandwidth, point estimates and robust confidence 
intervals for the RD estimand, when the outcome variable is categorical. The methods use
a nonparametric multinomial logit (MNL) transformation.
}
\details{
The details can be found in Xu (2017+). Introductions to regression discontinuity
designs can be found in Imbens and Lemieux (2008) and Cattaneo and Escanciano (2017).

[1] Cattaneo, M.D., and J.C. Escanciano (2017): "Introduction: Regression Discontinuity Designs,"
in Regression Discontinuity Designs: Theory and Applications, \emph{Advances in Econometrics}, volume 38.

[2] Imbens, G.W., and T. Lemieux (2008): "Regression Discontinuity Designs: a Guide to 
Practice," \emph{Journal of Econometrics}, 142, 615-635.

[3] Xu, K.-L. (2017+): "Regression Discontinuity with Categorical Outcomes," \emph{Journal of Econometrics}, in press.
}
\examples{
#Read your data into R, naming it as DAT (see above for the format of DAT)
#To illustrate, in the following I use the data in birth.weight.data (40750 obs)

>library("rd.categorical")
>data("birth.weight.data")
>DAT=birth.weight.data
>J=length(DAT[1,])-2
>c=1.5             #your cutoff value
>g=rd.mnl(DAT, c, matrix(0,J,1),diag(J),matrix(0,J,1),0.95)

>g$h_opt           #Chosen bandwidth
[1] 0.1788468

>g$ATE             #RD estimates
           [,1]
[1,] -0.01844537
[2,]  0.02043593

>g$ttest_rob       #95\% Robust t-tests
         [,1]         [,2]
[1,] -4.279627 1.872071e-05
[2,]  3.879861 1.045162e-04

>g$Waldtest_rob    #95\% Robust Wald joint-significance test
         [,1]        [,2]
[1,]  32.87078 7.28116e-08

}
\author{
Ke-Li Xu, Indiana University, \email{keli.xu15@gmail.com}
}
\keyword{RD}
