% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birth.weight.data.R
\docType{data}
\name{birth.weight.data}
\alias{birth.weight.data}
\title{A simulated dataset}
\description{
A simulated dataset (mimicking the California infant birth-weight example)
}
\details{
The dataset contains 40750 observations. The first column, bweight, is the running variable.
Columns 2-4 represent the 3 categories of the dependent variable (infant mortality), 
i.e. death within one-day, death between one-day and one-year, and alive.
}
\references{
The dataset is a realization of the DGP 1 in Xu (2017+, the Simulations 
subsection in Section 6.1)

Xu, K.-L. (2017+): "Regression Discontinuity with Categorical Outcomes," \emph{Journal of Econometrics}, in press.
}
\keyword{data}
